Renamed types -> common
