+++ ts/common.ts
