+++ ts/profile/EmblemImage.ts
@@ -1,4 +1,4 @@
-import {url} from '../types';
+import {url} from '../common';
 
 /**
  * The URL of an image of the player's emblem
